% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{add_validators}
\alias{add_validators}
\title{Add a validator to an existing object initialize function}
\usage{
add_validators(parent_initialize, validator_function)
}
\arguments{
\item{parent_initialize}{\code{derived_file}}

\item{validator_function}{Function which takes design as an argument}
}
\description{
Add a validator to an existing object initialize function
}
