% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\docType{data}
\name{Call}
\alias{Call}
\alias{rscript_call}
\alias{docker_call}
\title{A call}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Call

rscript_call(id, ..., design = NULL)

docker_call(id, ..., design = NULL)
}
\arguments{
\item{id}{The name of the call}

\item{...}{Extra arguments to the call, such as inputs, outputs, script, ...}

\item{design}{A dataframe containing the information for each call in a separate row}
}
\description{
A call
}
\keyword{datasets}
