% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\docType{data}
\name{Environment}
\alias{Environment}
\alias{local_environment}
\alias{docker_environment}
\title{Environments}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Environment

local_environment(...)

docker_environment(...)
}
\arguments{
\item{...}{Other parameters}
}
\description{
Environments
}
\keyword{datasets}
