% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executors.R
\docType{data}
\name{Executor}
\alias{Executor}
\alias{local_executor}
\alias{docker_executor}
\title{Executors}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Executor

local_executor(...)

docker_executor(...)
}
\arguments{
\item{...}{Other parameters}
}
\description{
Executors
}
\keyword{datasets}
