% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_set.R
\name{load_call}
\alias{load_call}
\title{Load a call from an R file, given by a "get_call" function in that R file}
\usage{
load_call(call_path, derived_file_directory = "./", ...)
}
\arguments{
\item{derived_file_directory}{Optional, the location in which derived files should be stored}

\item{...}{Other parameters given to the get_call function}

\item{call_location}{The location of the R file which, when sourced, contains a "get_call" function}
}
\description{
Load a call from an R file, given by a "get_call" function in that R file
}
