% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{Object}
\alias{Object}
\alias{docker}
\alias{derived_file}
\alias{raw_file}
\alias{script_file}
\alias{derived_directory}
\alias{parameters}
\alias{object_set}
\title{Objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Object

docker(...)

derived_file(...)

raw_file(...)

script_file(...)

derived_directory(...)

parameters(...)

object_set(...)
}
\arguments{
\item{...}{...}
}
\description{
Objects
}
\keyword{datasets}
