% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\docType{data}
\name{Workflow}
\alias{Workflow}
\alias{workflow}
\title{A workflow}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Workflow

workflow(...)
}
\arguments{
\item{...}{Extra arguments to the call, such as inputs, outputs, script, ...}
}
\description{
A workflow
}
\keyword{datasets}
