% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\docType{data}
\name{Call}
\alias{Call}
\alias{rscript_call}
\alias{rmd_call}
\title{A call}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Call

rscript_call(id = "", design, inputs, outputs)

rmd_call(id = "", design, inputs, outputs)
}
\arguments{
\item{id}{The name of the call}

\item{design}{The design}

\item{inputs}{The inputs}

\item{outputs}{The outputs}
}
\description{
A call
}
\keyword{datasets}
