% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{Object}
\alias{Object}
\alias{docker_container}
\alias{derived_file}
\alias{raw_file}
\alias{script_file}
\title{Objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Object

docker_container(...)

derived_file(...)

raw_file(...)

script_file(...)
}
\description{
Objects
}
\keyword{datasets}
