% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\docType{data}
\name{Workflow}
\alias{Workflow}
\title{A certigo workflow}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Workflow
}
\description{
A certigo workflow
}
\keyword{datasets}
